docs-pipeline/requirements.txt
requests>=2.31.0
pyyaml>=6.0
html2text>=2024.2.26
